\name{example1.member}
\docType{data}
\alias{example1.member}
\title{simulated SD1 containing 9 datasets, used in Figure 2B}
\description{nothing}
\usage{data(example1.member)}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{gene membership in example1 data}


