\name{example1}
\docType{data}
\alias{example1}
\title{simulated SD1 containing 9 datasets, used in Figure 2B}
\description{nothing}
\usage{data(example1)}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{datasets}


