\name{example2.member}
\docType{data}
\alias{example2.member}
\title{simulated SD1 containing 9 datasets, used in Figure 2B}
\description{nothing}
\usage{data(example2.member)}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{gene membership in example2 data}


