\name{examples.member}
\docType{data}
\alias{examples.member}
\title{simulated SD1 containing 9 datasets, used in Figure 2B}
\description{nothing}
\usage{data(examples.member)}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{gene membership in examples data}


