\name{examples}
\docType{data}
\alias{examples}
\title{100 trial datasets}
\description{100  simmulated datasets as shown in figure 2A}
\usage{data(examples)}
\details{
nothing
}
\value{
no return
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{datasets}
