\name{Multi_NB}
\alias{Multi_NB}
\title{main function in scLM}
\description{
An automatic detection algorithm for co-expressed genes across multiple single cell RNA-Seq (scRNA-seq) datasets
scLM is capable of identify co-expressed genes profiled by scRNA-Seq across multiple conditions without prior knowledge
}
\details{
nothing
}
\value{
the latent variables; coefficients; co-expression clusters; BIC values
}
\references{
nothing
}
\author{
Qianqian Song <qsong@wakehealth.edu>
}
\keyword{ scLM }

